fix: Promote Adobe's STAT_strings to universal, move current to GF
