+++ profile-adobe/src/checks/adobefonts/mod.rs
@@ -1,6 +1,3 @@
-#[allow(non_snake_case)]
-mod STAT_strings;
-pub use STAT_strings::STAT_strings;
 mod unsupported_tables;
 pub use unsupported_tables::unsupported_tables;
 mod nameid_1_win_english;
