+++ profile-adobe/src/lib.rs
@@ -9,12 +9,12 @@ impl fontspector_checkapi::ProfileProvider for Adobe {
     fn register(&self, cr: &mut Registry) -> Result<(), FontspectorError> {
         let builder = ProfileBuilder::new()
             .include_profile("universal")
+            // The universal profile now includes the "Adobe version" STAT_strings check, so we no longer need to add it here.
             .add_section("Adobe Fonts Checks")
             // "Adobe Fonts Checks" = ["adobefonts/family/consistent_upm", "adobefonts/nameid_1_win_english", "adobefonts/unsupported_tables", "adobefonts/STAT_strings"]
             .add_and_register_check(checks::adobefonts::family::consistent_upm)
             .add_and_register_check(checks::adobefonts::nameid_1_win_english)
             .add_and_register_check(checks::adobefonts::unsupported_tables)
-            .add_and_register_check(checks::adobefonts::STAT_strings)
             .exclude_check("opentype/xavgcharwidth")
             .exclude_check("designspace_has_consistent_codepoints")
             .exclude_check("designspace_has_consistent_glyphset")
@@ -27,7 +27,6 @@ impl fontspector_checkapi::ProfileProvider for Adobe {
             .exclude_check("ufo_recommended_fields")
             .exclude_check("ufo_required_fields")
             .exclude_check("ufo_unnecessary_fields")
-            .exclude_check("STAT_strings")
             .exclude_check("transformed_components")
             .exclude_check("unreachable_glyphs")
             .exclude_check("whitespace_ink")
