+++ profile-googlefonts/src/lib.rs
@@ -40,7 +40,9 @@ impl fontspector_checkapi::ProfileProvider for GoogleFonts {
         cr.register_filetype("IMAGE", IMAGE);
         cr.register_filetype("LICENSE", LICENSE);
 
-        let builder = ProfileBuilder::new().include_profile("universal");
+        let builder = ProfileBuilder::new()
+            .include_profile("universal")
+            .exclude_check("adobefonts/STAT_strings"); // We have our own stricter version
 
         #[cfg(feature = "check")]
         let builder = builder
@@ -51,30 +53,14 @@ impl fontspector_checkapi::ProfileProvider for GoogleFonts {
 
         #[cfg(feature = "check")]
         let builder = builder.add_and_register_check(checks::googlefonts::metadata::axes);
-        //            checks::googlefonts::metadata::axisregistry_bounds // Merged into metadata/axes
-        //            checks::googlefonts::metadata::axisregistry_valid_tags // Merged into metadata/axes
-        //            checks::googlefonts::metadata::consistent_axis_enumeration // Merged into metadata/axes
         #[cfg(all(feature = "check", not(target_family = "wasm")))]
         let builder = builder.add_and_register_check(checks::googlefonts::metadata::broken_links);
         #[cfg(feature = "check")]
         let builder = builder
-            //            checks::googlefonts::metadata::canonical_weight_value // Merged into metadata/validate
-            //            checks::googlefonts::metadata::designer_values // Merged into metadata/validate
-            //            checks::googlefonts::metadata::empty_designer // Merged into metadata/validate
             .add_and_register_check(checks::googlefonts::metadata::can_render_samples)
             .add_and_register_check(checks::googlefonts::metadata::category)
-            //            checks::googlefonts::metadata::category_hints // merged into metadata/validate
             .add_and_register_check(checks::googlefonts::metadata::consistent_repo_urls)
             .add_and_register_check(checks::googlefonts::metadata::consistent_with_fonts);
-        //            checks::googlefonts::metadata::filenames // merged into metadata/consistent_with_fonts
-        //            checks::googlefonts::metadata::canonical_style_names // merged into metadata/consistent_with_fonts
-        //            checks::googlefonts::metadata::valid_full_name_values // merged into metadata/consistent_with_fonts
-        //            checks::googlefonts::metadata::nameid/post_script_name // merged into metadata/consistent_with_fonts
-        //            checks::googlefonts::metadata::valid_post_script_name_values // merged into metadata/consistent_with_fonts
-        //            checks::googlefonts::metadata::nameid/family_and_full_names // merged into metadata/consistent_with_fonts
-        //            checks::googlefonts::metadata::valid_filename_values // redundant, see fontbakery#4997
-        //            checks::googlefonts::metadata::undeclared_fonts // redundant, see fontbakery#4997
-        //            checks::googlefonts::metadata::nameid/font_name // redundant, see fontbakery#4581
         #[cfg(all(feature = "check", not(target_family = "wasm")))]
         let builder =
             builder.add_and_register_check(checks::googlefonts::metadata::designer_profiles);
@@ -85,18 +71,10 @@ impl fontspector_checkapi::ProfileProvider for GoogleFonts {
             .add_and_register_check(checks::googlefonts::metadata::family_directory_name)
             .add_and_register_check(checks::googlefonts::metadata::familyname)
             .add_and_register_check(checks::googlefonts::metadata::has_regular)
-            //            checks::googlefonts::metadata::match_filename_postscript // Merged into metadata/validate
-            //            checks::googlefonts::metadata::match_fullname_postscript // Merged into metadata/validate
-            //            checks::googlefonts::metadata::match_name_familyname // Merged into metadata/validate
-            //            checks::googlefonts::metadata::match_weight_postscript // Merged into metadata/validate
-            //            checks::googlefonts::metadata::minisite_url // Merged into metadata/validate
-            //            checks::googlefonts::metadata::unique_full_name_values // Merged into metadata/validate
-            //            checks::googlefonts::metadata::unique_weight_style_pairs // Merged into metadata/validate
             .add_and_register_check(checks::googlefonts::metadata::primary_script)
             .add_and_register_check(checks::googlefonts::metadata::valid_primary_script_language)
             .add_and_register_check(checks::googlefonts::metadata::regular_is_400)
             .add_and_register_check(checks::googlefonts::metadata::subsets_correct) // Replacement for metadata/unsupported_subsets
-            //            checks::googlefonts::metadata::single_cjk_subset // Merged into metadata/subsets_correct
             .add_and_register_check(checks::googlefonts::metadata::unreachable_subsetting)
             .add_and_register_check(checks::googlefonts::metadata::validate)
             .add_and_register_check(checks::googlefonts::metadata::valid_nameid25)
@@ -159,10 +137,6 @@ impl fontspector_checkapi::ProfileProvider for GoogleFonts {
             .add_and_register_check(checks::googlefonts::repo::ascii_filenames)
             .add_and_register_check(checks::googlefonts::repo::dirname_matches_nameid_1)
             .add_and_register_check(checks::googlefonts::repo::vf_has_static_fonts)
-            //            checks::googlefonts::repo::fb_report // Upstream repos should be checked separately
-            //            checks::googlefonts::repo::sample_image // Upstream repos should be checked separately
-            //            checks::googlefonts::repo::upstream_yaml_has_required_fields // Redundant, no upstream.yaml any more
-            //            checks::googlefonts::repo::zip_files // Upstream repos should be checked separately
             .add_section("Shaping Checks")
             .add_and_register_check(checks::dotted_circle);
 
@@ -216,17 +190,14 @@ impl fontspector_checkapi::ProfileProvider for GoogleFonts {
             .add_and_register_check(checks::googlefonts::name::illegal_particles)
             .add_and_register_check(checks::googlefonts::name::version_format)
             .add_and_register_check(checks::googlefonts::parametric_axes_hidden)
-            // checks::googlefonts::production_encoded_glyphs // DISABLED
-            // checks::googlefonts::production_glyphs_similarity // Unlikely to be useful in the short term
-            // checks::googlefonts::description::family_update // Unlikely to useful yet
             .add_and_register_check(checks::googlefonts::render_own_name)
             .add_and_register_check(checks::googlefonts::STAT::axis_order)
             .add_and_register_check(checks::googlefonts::STAT::axisregistry)
             .add_and_register_check(checks::googlefonts::STAT::compulsory_axis_values)
             .add_and_register_check(checks::googlefonts::STAT::opsz_not_elided)
+            .add_and_register_check(checks::STAT_strings)
             .add_and_register_check(checks::googlefonts::unitsperem)
             .add_and_register_check(checks::googlefonts::use_typo_metrics)
-            // Not porting generate_static, see fontbakery#1727
             .add_and_register_check(checks::googlefonts::varfont::has_HVAR)
             .add_and_register_check(checks::googlefonts::varfont::slnt_needs_italic)
             .add_and_register_check(checks::googlefonts::vendor_id)
