+++ profile-universal/src/checks/STAT_strings.rs
@@ -5,10 +5,14 @@ use itertools::Itertools;
 use skrifa::raw::{tables::stat::AxisValue, TableProvider};
 
 #[check(
-    id = "STAT_strings",
+    id = "adobefonts/STAT_strings",
     rationale = "
-        On the STAT table, the \"Italic\" keyword must not be used on AxisValues
-        for variation axes other than 'ital'.
+        
+        In the STAT table, the \"Italic\" keyword must not be used on AxisValues
+        for variation axes other than 'ital' or 'slnt'. This is a more lenient
+        implementation of googlefonts/STAT_strings which allows \"Italic\"
+        only for the 'ital' axis.
+    
     ",
     proposal = "https://github.com/fonttools/fontbakery/issues/2863",
     title = "Check correctness of STAT table strings"
@@ -22,28 +26,34 @@ fn STAT_strings(t: &Testable, _context: &Context) -> CheckFnResult {
         .iter()
         .position(|axis| axis.axis_tag() == "ital")
         .map(|x| x as u16);
+    let slnt_pos = axes
+        .iter()
+        .position(|axis| axis.axis_tag() == "slnt")
+        .map(|x| x as u16);
     let mut name_ids = HashSet::new();
     if let Some(Ok(subtable)) = stat.offset_to_axis_values() {
         for axis_value in subtable.axis_values().iter().flatten() {
             match axis_value {
                 AxisValue::Format1(v) => {
-                    if Some(v.axis_index()) != ital_pos {
+                    if Some(v.axis_index()) != ital_pos && Some(v.axis_index()) != slnt_pos {
                         name_ids.insert(v.value_name_id());
                     }
                 }
                 AxisValue::Format2(v) => {
-                    if Some(v.axis_index()) != ital_pos {
+                    if Some(v.axis_index()) != ital_pos && Some(v.axis_index()) != slnt_pos {
                         name_ids.insert(v.value_name_id());
                     }
                 }
                 AxisValue::Format3(v) => {
-                    if Some(v.axis_index()) != ital_pos {
+                    if Some(v.axis_index()) != ital_pos && Some(v.axis_index()) != slnt_pos {
                         name_ids.insert(v.value_name_id());
                     }
                 }
                 AxisValue::Format4(v) => {
                     for axis_value in v.axis_values() {
-                        if Some(axis_value.axis_index()) != ital_pos {
+                        if Some(axis_value.axis_index()) != ital_pos
+                            && Some(axis_value.axis_index()) != slnt_pos
+                        {
                             name_ids.insert(v.value_name_id());
                         }
                     }
@@ -67,25 +77,5 @@ fn STAT_strings(t: &Testable, _context: &Context) -> CheckFnResult {
         ))
     } else {
         Ok(Status::just_one_pass())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use fontspector_checkapi::codetesting::{
-        assert_pass, assert_results_contain, run_check, test_able,
-    };
-
-    use fontspector_checkapi::StatusCode;
-
-    #[test]
-    fn test_stat_strings() {
-        let testable = test_able("ibmplexsans-vf/IBMPlexSansVar-Roman.ttf");
-        let results = run_check(super::STAT_strings, testable);
-        assert_pass(&results);
-
-        let testable = test_able("ibmplexsans-vf/IBMPlexSansVar-Italic.ttf");
-        let results = run_check(super::STAT_strings, testable);
-        assert_results_contain(&results, StatusCode::Fail, Some("bad-italic".to_string()));
     }
 }
